Refactor MQTT Broker Tests and Update Build Configuration
- Updated MQTT broker tests to use new class structure and methods.
- Changed MQTT configuration structure to align with new implementation.
- Enhanced client management tests to include client information retrieval.
- Added tests for message publishing and retained messages.
- Updated XMake build configuration to include necessary dependencies and paths.
- Added placeholder tests for comprehensive, integration, performance, and protocol testing.
- Introduced a stub implementation for MovableBehavior to facilitate compilation.

+++ tests/framework/comprehensive_test_framework.cpp
@@ -292,10 +292,10 @@ json ComprehensiveTestFixture::generateTestData(const std::string& schema) {
 
 std::vector<uint8_t> ComprehensiveTestFixture::generateRandomData(size_t size) {
     std::vector<uint8_t> data(size);
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
-    
+    std::uniform_int_distribution<int> dist(0, 255);
+
     for (auto& byte : data) {
-        byte = dist(randomGenerator_);
+        byte = static_cast<uint8_t>(dist(randomGenerator_));
     }
     
     return data;
